test(test-optimization): run worker test in Jest matrix
